requirements.txt
SpeechRecognition
PyAudio
gTTS
google-generativeai
openai
python-dotenv
pygame
pyttsx3
